--- packages/logger/src/client/.env.schema
@@ -1,4 +1,0 @@
-NODE_ENV=development
-TEST_MODE=false
-# Log level is only applied in production mode, in development logs will always be debug level
-# LOG_LEVEL=info
