+++ packages/logger/src/client/stringUtils.ts
